// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/changeUserPasswordInput.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * The version of the OpenAPI document: 3.13.0-SNAPSHOT
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChangeUserPasswordInput { 
    /**
     * Token of the user to be validated.
     */
    token: string;
    /**
     * Password of the user.
     */
    password: string;
    /**
     * First name of the user.
     */
    firstname: string;
    /**
     * Last name of the user.
     */
    lastname: string;
}